Spring Initializr demonstration
Let's create a new project right in Intellij IDEA using the New Project
wizard for a rest/jpa secured Spring Boot app using the actuator.

The wizard automatically creates a fileset including the configuration
and our main app. It also creates a run configuration for the app so
that we can directly start it.

Before doing so, let's create a simple entity with its (rest) repository
and an `import.sql` that insert some entries on startup.