Manual meta-data demonstration (I)
We can fix the warning for our custom `foo.counter` via a quick fix to
create a meta-data entry manually.

We get content assistance in the json document and syntax highlighting
for description and values. We also get a warning if the description does
not end with a dot.

We can also create an extra property to show the assistance.

The banner tells us that the meta-data processor is not available in the
project so we can add the dependency and rebuild the project. Our two
custom keys are recognized.