Manual meta-data demonstration (II)
Spring Boot 1.3 introduced the notion of value hints. Let's specify that
that counter can take two values (42 and 7) with a relevant description.

Our `foo.counter` can only have the value of 42 and 7. What if these were
only hints and any extra value is allowed. We can add a value provider
for that.

IJ provides auto-completion for all the available providers. More can be
added in future release. Let's update our `foo.bar` to actually represent
the name of a Spring Bean. Create two spring beans of whatever type, for
instance String, and restrict the beans of String type in the context.